+++ CHANGELOG.md
@@ -9,4 +9,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 
+- Installing mods for a Fabric server from Modrinth not working
+
 ### Misc